fixed scroll script location
