+++ react-app/public/index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>JumpStart</title>
     <script src="https://kit.fontawesome.com/08886d62ae.js" crossorigin="anonymous"></script>
-    <script src="scrollScript.js"></script>
   </head>
   <body>
     <div id="root"></div>
