+++ react-app/src/components/HomePage/homePage.js
@@ -1,9 +1,44 @@
 import React, { useEffect } from "react";
 import { useSelector, useDispatch } from 'react-redux';
 import { Link } from "react-router-dom";
-import { getHomePageProjects } from '../../store/project.js'
-import csc from "country-state-city"
-import "./homePage.css"
+import { getHomePageProjects } from '../../store/project.js';
+import csc from "country-state-city";
+import "./homePage.css";
+
+// The debounce function receives our function as a parameter
+const debounce = (fn) => {
+
+    // This holds the requestAnimationFrame reference, so we can cancel it if we wish
+    let frame;
+
+    // The debounce function returns a new function that can receive a variable number of arguments
+    return (...params) => {
+
+      // If the frame variable has been defined, clear it now, and queue for next frame
+      if (frame) {
+        cancelAnimationFrame(frame);
+      }
+
+      // Queue our function call for the next frame
+      frame = requestAnimationFrame(() => {
+
+        // Call our function and pass any params we received
+        fn(...params);
+      });
+
+    }
+  }
+  // Reads out the scroll position and stores it in the data attribute
+  // so we can use it in our stylesheets
+  const storeScroll = () => {
+    document.documentElement.dataset.scroll = window.scrollY;
+  }
+
+  // Listen for new scroll events, here we debounce our `storeScroll` function
+  document.addEventListener('scroll', debounce(storeScroll), { passive: true });
+
+  // Update scroll position for first time
+  storeScroll();
 
 
 const HomePage = () => {
@@ -80,7 +115,7 @@ const HomePage = () => {
                                         <div id="meter">
                                             <span id="progressBar" style={{width: `${getPercentage(project)}%`}}></span>
                                         </div>
-                                        <div id="projectCar-amount">{`$${getSum(project)} raised out of $${project.goalAmount}`}</div>
+                                        <div id="projectCard-amount">{`$${getSum(project)} raised out of $${project.goalAmount}`}</div>
                                     </div>
                                 </Link>
                             ))}
@@ -103,7 +138,7 @@ const HomePage = () => {
                                         <div id="meter">
                                             <span id="progressBar" style={{width: `${getPercentage(project)}%`}}></span>
                                         </div>
-                                        <div id="projectCar-amount">{`$${getSum(project)} raised out of $${project.goalAmount}`}</div>
+                                        <div id="projectCard-amount">{`$${getSum(project)} raised out of $${project.goalAmount}`}</div>
                                     </div>
                                 </Link>
                             ))}
@@ -126,7 +161,7 @@ const HomePage = () => {
                                         <div id="meter">
                                             <span id="progressBar" style={{width: `${getPercentage(project)}%`}}></span>
                                         </div>
-                                        <div id="projectCar-amount">{`$${getSum(project)} raised out of $${project.goalAmount}`}</div>
+                                        <div id="projectCard-amount">{`$${getSum(project)} raised out of $${project.goalAmount}`}</div>
                                     </div>
                                 </Link>
                             ))}
